# KiemTraThichUng.API/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "KiemTraThichUng API",
    "description": "API documentation for Adaptive Testing System",
    "version": "v1"
  },
  "paths": {
...
    "/api/DataControls/GetBoCauHoi": {
      "post": {
        "tags": [
          "DataControls"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetAllBoCauHoi.GetAllBoCauHoiQuery"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GetAllBoCauHoi.GetAllBoCauHoiQuery"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GetAllBoCauHoi.GetAllBoCauHoiQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.IReadOnlyList.DataControls.BoCauHoiItemResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.IReadOnlyList.DataControls.BoCauHoiItemResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.IReadOnlyList.DataControls.BoCauHoiItemResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/DataControls/GetCauTrucBch": {
      "post": {
        "tags": [
          "DataControls"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetAllCauTruc.GetAllCauTrucQuery"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GetAllCauTruc.GetAllCauTrucQuery"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GetAllCauTruc.GetAllCauTrucQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.IReadOnlyList.DataControls.CauTrucItemResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.IReadOnlyList.DataControls.CauTrucItemResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.IReadOnlyList.DataControls.CauTrucItemResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/DataControls/GetLoaiCauHoiForSelector": {
      "post": {
        "tags": [
          "DataControls"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetLoaiCauHoiForSelector.GetLoaiCauHoiForSelectorQuery"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GetLoaiCauHoiForSelector.GetLoaiCauHoiForSelectorQuery"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GetLoaiCauHoiForSelector.GetLoaiCauHoiForSelectorQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.List.DataControls.SelectorItemResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.List.DataControls.SelectorItemResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.List.DataControls.SelectorItemResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/DataControls/GetMucDoNhanThucForSelector": {
      "post": {
        "tags": [
          "DataControls"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetMucDoNhanThucForSelector.GetMucDoNhanThucForSelectorQuery"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GetMucDoNhanThucForSelector.GetMucDoNhanThucForSelectorQuery"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GetMucDoNhanThucForSelector.GetMucDoNhanThucForSelectorQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.List.DataControls.SelectorItemResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.List.DataControls.SelectorItemResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.List.DataControls.SelectorItemResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/EX_CauHois/GetList": {
      "post": {
        "tags": [
          "EX_CauHois"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetListCauHoi.GetListCauHoiQuery"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GetListCauHoi.GetListCauHoiQuery"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GetListCauHoi.GetListCauHoiQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.PagedResult.EX_CauHois.CauHoiItemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.PagedResult.EX_CauHois.CauHoiItemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.PagedResult.EX_CauHois.CauHoiItemDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/EX_CauHois/GetById": {
      "post": {
        "tags": [
          "EX_CauHois"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetCauHoiById.GetCauHoiByIdQuery"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GetCauHoiById.GetCauHoiByIdQuery"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GetCauHoiById.GetCauHoiByIdQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.EX_CauHois.CauHoiDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.EX_CauHois.CauHoiDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.EX_CauHois.CauHoiDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/EX_CauHois": {
      "post": {
        "tags": [
          "EX_CauHois"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCauHoi.CreateCauHoiCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCauHoi.CreateCauHoiCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCauHoi.CreateCauHoiCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.CreateCauHoi.CreateCauHoiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.CreateCauHoi.CreateCauHoiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.CreateCauHoi.CreateCauHoiResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "EX_CauHois"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCauHoi.UpdateCauHoiCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCauHoi.UpdateCauHoiCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCauHoi.UpdateCauHoiCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.UpdateCauHoi.UpdateCauHoiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.UpdateCauHoi.UpdateCauHoiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.UpdateCauHoi.UpdateCauHoiResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "EX_CauHois"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteCauHoi.DeleteCauHoiCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteCauHoi.DeleteCauHoiCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteCauHoi.DeleteCauHoiCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.IReadOnlyList.DeleteCauHoi.DeleteCauHoiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.IReadOnlyList.DeleteCauHoi.DeleteCauHoiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.IReadOnlyList.DeleteCauHoi.DeleteCauHoiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/EX_CauHois/CapNhatTrangThaiCauHoi": {
      "put": {
        "tags": [
          "EX_CauHois"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTrangThaiCauHoi.CapNhatTrangThaiCauHoiCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTrangThaiCauHoi.CapNhatTrangThaiCauHoiCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTrangThaiCauHoi.CapNhatTrangThaiCauHoiCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.IReadOnlyList.UpdateTrangThaiCauHoi.CapNhatTrangThaiCauHoiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.IReadOnlyList.UpdateTrangThaiCauHoi.CapNhatTrangThaiCauHoiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.IReadOnlyList.UpdateTrangThaiCauHoi.CapNhatTrangThaiCauHoiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/ExamSelection/GetList": {
      "post": {
        "tags": [
          "ExamSelection"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LayDanhSachDeKiemTra.GetListByIdParentQuery"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LayDanhSachDeKiemTra.GetListByIdParentQuery"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LayDanhSachDeKiemTra.GetListByIdParentQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.PagedResult.ExamSelection.DeKiemTraitemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.PagedResult.ExamSelection.DeKiemTraitemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.PagedResult.ExamSelection.DeKiemTraitemDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/ExamSelection/GetById": {
      "post": {
        "tags": [
          "ExamSelection"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LayDeKiemTraById.LayDeKiemTraByIdQuery"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LayDeKiemTraById.LayDeKiemTraByIdQuery"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LayDeKiemTraById.LayDeKiemTraByIdQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.ExamSelection.DeKiemTraDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.ExamSelection.DeKiemTraDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.ExamSelection.DeKiemTraDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApiResponse.CreateCauHoi.CreateCauHoiResponse": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/CreateCauHoi.CreateCauHoiResponse"
          },
          "warningMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Responses.ApiMessage"
            },
            "nullable": true,
            "readOnly": true
          },
          "errorMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Responses.ApiMessage"
            },
            "nullable": true,
            "readOnly": true
          },
          "isOk": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "ApiResponse.EX_CauHois.CauHoiDto": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/EX_CauHois.CauHoiDto"
          },
          "warningMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Responses.ApiMessage"
            },
            "nullable": true,
            "readOnly": true
          },
          "errorMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Responses.ApiMessage"
            },
            "nullable": true,
            "readOnly": true
          },
          "isOk": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "ApiResponse.IReadOnlyList.DataControls.BoCauHoiItemResponse": {
        "type": "object",
        "properties": {
          "result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataControls.BoCauHoiItemResponse"
            },
            "nullable": true,
            "readOnly": true
          },
          "warningMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Responses.ApiMessage"
            },
            "nullable": true,
            "readOnly": true
          },
          "errorMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Responses.ApiMessage"
            },
            "nullable": true,
            "readOnly": true
          },
          "isOk": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "ApiResponse.IReadOnlyList.DataControls.CauTrucItemResponse": {
        "type": "object",
        "properties": {
          "result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataControls.CauTrucItemResponse"
            },
            "nullable": true,
            "readOnly": true
          },
          "warningMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Responses.ApiMessage"
            },
            "nullable": true,
            "readOnly": true
          },
          "errorMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Responses.ApiMessage"
            },
            "nullable": true,
            "readOnly": true
          },
          "isOk": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "ApiResponse.List.DataControls.SelectorItemResponse": {
        "type": "object",
        "properties": {
          "result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataControls.SelectorItemResponse"
            },
            "nullable": true,
            "readOnly": true
          },
          "warningMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Responses.ApiMessage"
            },
            "nullable": true,
            "readOnly": true
          },
          "errorMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Responses.ApiMessage"
            },
            "nullable": true,
            "readOnly": true
          },
          "isOk": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "ApiResponse.PagedResult.EX_CauHois.CauHoiItemDto": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/PagedResult.EX_CauHois.CauHoiItemDto"
          },
          "warningMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Responses.ApiMessage"
            },
            "nullable": true,
            "readOnly": true
          },
          "errorMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Responses.ApiMessage"
            },
            "nullable": true,
            "readOnly": true
          },
          "isOk": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "CreateCauHoi.CreateCauHoiCommand": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "maCauHoi": {
            "type": "string",
            "nullable": true
          },
          "idCauTruc": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "idLoaiCauHoi": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "idMucDoNhanThuc": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "idTrangThai": {
            "$ref": "#/components/schemas/Enums.TrangThaiDuyet"
          },
          "isKhongDao": {
            "type": "boolean",
            "nullable": true
          },
          "isCauHoiCha": {
            "type": "boolean",
            "nullable": true
          },
          "doKho": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "doKhoKhoiTao": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "noiDung": {
            "type": "string",
            "nullable": true
          },
          "stt": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "ghiChu": {
            "type": "string",
            "nullable": true
          },
          "mediaUrl": {
            "type": "string",
            "nullable": true
          },
          "tieuDeVeTrai": {
            "type": "string",
            "nullable": true
          },
          "tieuDeVePhai": {
            "type": "string",
            "nullable": true
          },
          "cauHois": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EX_CauHois.CauHoisDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DeleteCauHoi.DeleteCauHoiCommand": {
        "type": "object",
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Enums.TrangThaiDuyet": {
        "enum": [
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "GetAllBoCauHoi.GetAllBoCauHoiQuery": {
        "type": "object",
        "properties": {
          "pageNumber": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "pageSize": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "idParent": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "keyword": {
            "type": "string",
            "nullable": true
          },
          "sortCol": {
            "type": "string",
            "nullable": true
          },
          "isAsc": {
            "type": "boolean",
            "nullable": true
          },
          "isVisible": {
            "type": "boolean",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GetAllCauTruc.GetAllCauTrucQuery": {
        "type": "object",
        "properties": {
          "idBoCauHoi": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "idParent": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "keyword": {
            "type": "string",
            "nullable": true
          },
          "sortCol": {
            "type": "string",
            "nullable": true
          },
          "isAsc": {
            "type": "boolean"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "GetCauHoiById.GetCauHoiByIdQuery": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "GetListCauHoi.GetListCauHoiQuery": {
        "type": "object",
        "properties": {
          "pageNumber": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "pageSize": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "keyword": {
            "type": "string",
            "nullable": true
          },
          "sortCol": {
            "type": "string",
            "nullable": true
          },
          "isAsc": {
            "type": "boolean",
            "nullable": true
          },
          "maCauHoi": {
            "type": "string",
            "nullable": true
          },
          "isVisible": {
            "type": "boolean",
            "nullable": true
          },
          "tuNgayTao": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "denNgayTao": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "idMucDoNhanThuc": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "idTrangThaiCauHoi": {
            "$ref": "#/components/schemas/Enums.TrangThaiDuyet"
          },
          "idLoaiCauHoi": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "idCauTruc": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "idBoCauHoi": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "idNguoiSoan": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "idNhanSu": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GetLoaiCauHoiForSelector.GetLoaiCauHoiForSelectorQuery": {
        "type": "object",
        "properties": {
          "isVisible": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "GetMucDoNhanThucForSelector.GetMucDoNhanThucForSelectorQuery": {
        "type": "object",
        "properties": {
          "isVisible": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Format: Bearer {your token}",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}